feat : "User 검증방식 변경"

+++ src/main/java/com/side/football_project/reservation/service/ReservationServiceImpl.java
@@ -1,36 +1,40 @@
 package com.side.football_project.reservation.service;
 
-import com.side.football_project.domain.user.dto.UserResponseDto;
+import com.side.football_project.domain.stadium.entity.Stadium;
+import com.side.football_project.domain.stadium.service.StadiumService;
+import com.side.football_project.domain.user.entity.User;
 import com.side.football_project.domain.user.service.UserService;
-import com.side.football_project.global.security.auth.SessionAuthenticationService;
+import com.side.football_project.global.common.exception.CustomException;
+import com.side.football_project.global.common.exception.type.ReservationErrorCode;
+import com.side.football_project.global.common.exception.type.UserErrorCode;
 import com.side.football_project.reservation.domain.Reservation;
+import com.side.football_project.reservation.dto.ReservationDeleteRequestDto;
 import com.side.football_project.reservation.dto.ReservationRequestDto;
 import com.side.football_project.reservation.dto.ReservationResponseDto;
 import com.side.football_project.reservation.repository.ReservationRepository;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ReservationServiceImpl implements ReservationService {
 
     private final ReservationRepository reservationRepository;
     private final UserService userService;
-    private final SessionAuthenticationService sessionAuthenticationService;
     private final StadiumService stadiumService;
 
     @Transactional
     @Override
-    public ReservationResponseDto createReservation(ReservationRequestDto requestDto, HttpServletRequest request) {
+    public ReservationResponseDto createReservation(ReservationRequestDto requestDto) {
 
-        sessionAuthenticationService.authenticate(requestDto.getUser().getEmail(), requestDto.getUser().getPassword(), request);
-
-        UserResponseDto user = userService.findUser(requestDto.getUser().getId());
+        User user = userService.getLoginUser();
+        User findUser = userService.getUserFromDB(user.getId());
         Stadium stadium = stadiumService.findByIdWithLock(requestDto.getStadium().getId());
 
         if (stadium.isFullyBooked()) {
@@ -40,7 +44,7 @@ public class ReservationServiceImpl implements ReservationService {
         Reservation reservation = Reservation.builder()
                 .name(requestDto.getName())
                 .fee(requestDto.getFee())
-                .user(UserResponseDto.toEntity(user))
+                .user(findUser)
                 .stadium(stadium)
                 .build();
 
@@ -66,23 +70,29 @@ public class ReservationServiceImpl implements ReservationService {
     @Override
     public ReservationResponseDto findReservationByStadium(Long stadiumId) {
         Reservation reservation = reservationRepository.findReservationByStadiumId(stadiumId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 경기장에 예약된 내역이 없습니다. id=" + stadiumId));
+                .orElseThrow(() -> new CustomException(ReservationErrorCode.RESERVATION_NOT_FOUND));
 
         return ReservationResponseDto.toEntity(reservation);
     }
 
     @Override
-    public Page<ReservationResponseDto> findReservationByUser(Long id, int page, int size) {
+    public Page<ReservationResponseDto> findReservationByUser(Long userId, int page, int size) {
         Pageable pageable = PageRequest.of(page, size);
-        Page<Reservation> reservation = reservationRepository.findReservationByUser(id, pageable);
+        Page<Reservation> reservation = reservationRepository.findReservationByUser(userId, pageable);
         return reservation.map(ReservationResponseDto::toEntity);
     }
 
     @Transactional
     @Override
-    public void deleteReservation(Long id) {
-        Reservation reservation = reservationRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("해당 예약이 없습니다. id=" + id));
+    public void deleteReservation(ReservationDeleteRequestDto requestDto) {
+        User user = userService.getLoginUser();
+
+        Reservation reservation = reservationRepository.findById(requestDto.getId())
+                .orElseThrow(() -> new IllegalArgumentException("해당 예약이 없습니다. id=" + requestDto.getId()));
+
+        if (!reservation.getUser().getId().equals(user.getId())) {
+            throw new CustomException(UserErrorCode.USER_NOT_MATCH);
+        }
 
         reservationRepository.delete(reservation);
     }
